Add web worker for graph layout calculation
1. Install workerrize-loader, redux-thunk and redux-watch
2. Add an SVG loading spinner
3. Subscribe to the store and dispatch an async action to calculate the graph when graph input props change

Note: Currently only works in standalone app, not in exported lib
