+++ package.json
@@ -48,6 +48,8 @@
     "react-scripts": "^3.4.1",
     "react-test-renderer": "^16.13.1",
     "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0",
+    "redux-watch": "^1.1.1",
     "reselect": "^4.0.0",
     "seedrandom": "^3.0.5",
     "snyk": "^1.341.1",
@@ -77,7 +79,8 @@
     "stylelint-config-prettier": "^8.0.1",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-order": "^4.0.0",
-    "stylelint-prettier": "^1.1.2"
+    "stylelint-prettier": "^1.1.2",
+    "workerize-loader": "^1.3.0"
   },
   "husky": {
     "hooks": {
