+++ src/components/icons/loading.scss
@@ -1,0 +1,62 @@
+@import '../../styles/variables';
+
+$polygon-length: 537.146;
+$section-length: $polygon-length / 6;
+$section-gap: $section-length / 6;
+$section-gap-2x: $section-gap * 2;
+$spin1-offset-start: -$section-gap;
+$spin1-offset-end: -$polygon-length - $section-gap;
+$spin2-offset-start: -$section-length - $section-gap-2x;
+$spin2-offset-end: -$polygon-length + -$section-length - $section-gap-2x;
+$path1-length: $section-length;
+$path1-gap: $polygon-length - $section-length;
+$path2-length: $polygon-length - $section-length - $section-gap-2x;
+$path2-gap: $section-length + $section-gap-2x;
+
+.pipeline-loading-icon {
+  display: block;
+  width: 30px;
+  height: auto;
+
+  path {
+    fill: none;
+    stroke-width: 10px;
+  }
+
+  path:nth-of-type(1) {
+    animation: stroke-spin1 2s infinite linear;
+    stroke: $color-primary;
+    stroke-dasharray: $path1-length, $path1-gap;
+  }
+
+  path:nth-of-type(2) {
+    animation: stroke-spin2 2s infinite linear;
+    stroke-dasharray: $path2-length, $path2-gap;
+
+    .kui-theme--light & {
+      stroke: $color-bg-light-alt;
+    }
+
+    .kui-theme--dark & {
+      stroke: $color-bg-dark-alt;
+    }
+  }
+}
+
+@keyframes stroke-spin1 {
+  from {
+    stroke-dashoffset: $spin1-offset-start;
+  }
+  to {
+    stroke-dashoffset: $spin1-offset-end;
+  }
+}
+
+@keyframes stroke-spin2 {
+  from {
+    stroke-dashoffset: $spin2-offset-start;
+  }
+  to {
+    stroke-dashoffset: $spin2-offset-end;
+  }
+}
