+++ src/components/wrapper/index.js
@@ -4,12 +4,13 @@ import classnames from 'classnames';
 import FlowChart from '../flowchart';
 import Sidebar from '../sidebar';
 import ExportModal from '../export-modal';
+import LoadingIcon from '../icons/loading';
 import './wrapper.css';
 
 /**
  * Main app container. Handles showing/hiding the sidebar nav, and theme classes.
  */
-export const Wrapper = ({ theme }) => (
+export const Wrapper = ({ graph, loading, theme }) => (
   <div
     className={classnames('kedro-pipeline', {
       'kui-theme--dark': theme === 'dark',
@@ -17,6 +18,7 @@ export const Wrapper = ({ theme }) => (
     })}>
     <Sidebar />
     <div className="pipeline-wrapper">
+      {loading && <LoadingIcon className="pipeline-wrapper__loading" />}
       <FlowChart />
     </div>
     <ExportModal />
@@ -24,6 +26,8 @@ export const Wrapper = ({ theme }) => (
 );
 
 export const mapStateToProps = state => ({
+  loading: state.loading,
+  graph: state.graph,
   theme: state.theme
 });
