+++ src/selectors/layout.js
@@ -1,12 +1,11 @@
 import { createSelector } from 'reselect';
-import dagre from 'dagre';
-import { graph } from '../utils/graph/';
 import { getCurrentFlags } from './flags';
 import { getVisibleNodes } from './nodes';
 import { getVisibleEdges } from './edges';
 import { getVisibleLayerIDs } from './disabled';
 import { sidebarBreakpoint, sidebarWidth } from '../config';
 
+const getGraph = state => state.graph;
 const getHasVisibleLayers = state =>
   state.visible.layers && Boolean(state.layer.ids.length);
 const getNodeType = state => state.node.type;
@@ -17,7 +16,7 @@ const getFontLoaded = state => state.fontLoaded;
 /**
  * Calculate chart layout. Algorithm used is dependent on flags
  */
-export const getGraph = createSelector(
+export const getGraphInput = createSelector(
   [
     getVisibleNodes,
     getVisibleEdges,
@@ -30,54 +29,9 @@ export const getGraph = createSelector(
     if (!fontLoaded || !nodes.length || !edges.length) {
       return;
     }
-
-    // Use experimental graph rendering if flag enabled
-    if (flags.newgraph) {
-      const result = graph(nodes, edges, showLayers && layers);
-
-      return {
-        graph: () => ({ ...result.size, marginx: 100, marginy: 100 }),
-        nodes: () => result.nodes.map(node => node.id),
-        edges: () => result.edges.map(edge => edge.id),
-        node: id => result.nodes.find(node => node.id === id),
-        edge: id => result.edges.find(edge => edge.id === id),
-        newgraph: true
-      };
-    }
-
-    // Otherwise use dagre to render
-    return graphDagre(nodes, edges, showLayers);
+    return { nodes, edges, layers, showLayers, flags, fontLoaded };
   }
 );
-
-/**
- * Calculate chart layout with Dagre.js.
- * This is an extremely expensive operation so we want it to run as infrequently
- * as possible, and keep it separate from other properties (like node.active)
- * which don't affect layout.
- */
-export const graphDagre = (nodes, edges, hasVisibleLayers) => {
-  const ranker = hasVisibleLayers ? 'none' : null;
-  const graph = new dagre.graphlib.Graph().setGraph({
-    ranker: hasVisibleLayers ? ranker : null,
-    ranksep: hasVisibleLayers ? 200 : 70,
-    marginx: 40,
-    marginy: 40
-  });
-
-  nodes.forEach(node => {
-    graph.setNode(node.id, node);
-  });
-
-  edges.forEach(edge => {
-    graph.setEdge(edge.source, edge.target, edge);
-  });
-
-  // Run Dagre layout to calculate X/Y positioning
-  dagre.layout(graph);
-
-  return graph;
-};
 
 /**
  * Reformat node data for use on the chart,
@@ -87,11 +41,10 @@ export const getLayoutNodes = createSelector(
   [getGraph, getNodeType, getNodeLayer],
   (graph, nodeType, nodeLayer) =>
     graph
-      ? graph.nodes().map(nodeID => {
-          const node = graph.node(nodeID);
+      ? graph.nodes.map(node => {
           return Object.assign({}, node, {
-            layer: nodeLayer[nodeID],
-            type: nodeType[nodeID],
+            layer: nodeLayer[node.id],
+            type: nodeType[node.id],
             order: node.x + node.y * 9999
           });
         })
@@ -103,8 +56,7 @@ export const getLayoutNodes = createSelector(
  */
 export const getLayoutEdges = createSelector(
   [getGraph],
-  graph =>
-    graph ? graph.edges().map(edge => Object.assign({}, graph.edge(edge))) : []
+  graph => (graph ? graph.edges : [])
 );
 
 /**
@@ -112,7 +64,7 @@ export const getLayoutEdges = createSelector(
  */
 export const getGraphSize = createSelector(
   [getGraph],
-  graph => (graph ? graph.graph() : {})
+  graph => (graph ? graph.graph : {})
 );
 
 /**
