+++ src/store/index.js
@@ -1,5 +1,9 @@
-import { createStore } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
+import thunk from 'redux-thunk';
+import watch from 'redux-watch';
 import reducer from '../reducers';
+import { getGraphInput } from '../selectors/layout';
+import { calculateGraph } from '../actions/graph';
 import { saveState } from './helpers';
 
 /**
@@ -8,7 +12,14 @@ import { saveState } from './helpers';
  * @return {Object} Redux store
  */
 export default function configureStore(initialState) {
-  const store = createStore(reducer, initialState);
+  const store = createStore(reducer, initialState, applyMiddleware(thunk));
+
+  let watchGraph = watch(() => getGraphInput(store.getState()));
+  store.subscribe(
+    watchGraph(graphInput => {
+      store.dispatch(calculateGraph(graphInput));
+    })
+  );
 
   store.subscribe(() => {
     const { textLabels, theme, nodeType, visible, flags } = store.getState();
