+++ src/utils/graph/worker.js
@@ -1,0 +1,3 @@
+import worker from 'workerize-loader!./index'; // eslint-disable-line import/no-webpack-loader-syntax
+
+export default worker;
